Add piwik tracking support
Piwik tracking is an analytics service akin to Google
Analytics (can be self-hosted). Support for this service,
while in the default theme, was missing from this theme.

I kept the same variables as the default theme that need
to be present in pelicanconf.py (or publishconf.py):

PIWIK_URL
PIWIK_SITE_ID

The default theme includes PIWIK_SSL_URL, but PIWIK seems
to recommend just using the universal // prefix, so I didn't
bother to include that. As such, users should use whatever
prefix they want (http://, //, https://).  This implementation
of Piwik's javascript snippet is merely a modification of the
one provided when you make a new site.