Make Sidebar use Bootstrap colouring
We used to manually provide the colour of the sidebar as well as the
'view all articles' button. This is a problem when using some bootstrap
colourschemes that conflict with the provided colour.

This removes the manual twitchy-background css class and uses
btn-default to style the backgrounds of the theme which should never
conflict with the themes.